# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CharacterOut": {
        "properties": {
          "alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alias"
          },
          "appearance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearance"
          },
          "background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background"
          },
          "contradiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contradiction"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "fatal_weakness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fatal Weakness"
          },
          "first_chapter_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Chapter Seq"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "role": {
            "default": "supporting",
            "enum": [
              "protagonist",
              "supporting",
              "antagonist",
              "minor"
            ],
            "title": "Role",
            "type": "string"
          },
          "secret_desire": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Desire"
          },
          "status": {
            "default": "alive",
            "enum": [
              "alive",
              "dead",
              "missing",
              "unknown"
            ],
            "title": "Status",
            "type": "string"
          },
          "surface_identity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Identity"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "CharacterOut",
        "type": "object"
      },
      "CharacterRelationInput": {
        "properties": {
          "from_char_id": {
            "title": "From Char Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "relation_type": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Relation Type",
            "type": "string"
          },
          "to_char_id": {
            "title": "To Char Id",
            "type": "integer"
          }
        },
        "required": [
          "from_char_id",
          "to_char_id",
          "relation_type"
        ],
        "title": "CharacterRelationInput",
        "type": "object"
      },
      "CharacterRelationOut": {
        "description": "人物关系边（关系图谱数据源）。",
        "properties": {
          "from_char_id": {
            "title": "From Char Id",
            "type": "integer"
          },
          "from_name": {
            "default": "",
            "title": "From Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "relation_type": {
            "title": "Relation Type",
            "type": "string"
          },
          "to_char_id": {
            "title": "To Char Id",
            "type": "integer"
          },
          "to_name": {
            "default": "",
            "title": "To Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "from_char_id",
          "to_char_id",
          "relation_type"
        ],
        "title": "CharacterRelationOut",
        "type": "object"
      },
...
      "ForeshadowOut": {
        "properties": {
          "actual_payoff_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Payoff Seq"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "importance": {
            "default": "medium",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Importance",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "planned_payoff_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Payoff Seq"
          },
          "planted_chapter_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planted Chapter Seq"
          },
          "status": {
            "default": "open",
            "enum": [
              "open",
              "closed",
              "abandoned"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "title",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "ForeshadowOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI 小说创作工具",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/books/{book}/character-relations": {
      "get": {
        "operationId": "list_character_relations_api_books__book__character_relations_get",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharacterRelationOut"
                  },
                  "title": "Response List Character Relations Api Books  Book  Character Relations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Character Relations",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_character_relation_api_books__book__character_relations_post",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterRelationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterRelationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Character Relation",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/books/{book}/character-relations/{relation_id}": {
      "delete": {
        "operationId": "delete_character_relation_api_books__book__character_relations__relation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          },
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "relation_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Relation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Character Relation",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/books/{book}/characters": {
      "get": {
        "operationId": "list_characters_api_books__book__characters_get",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CharacterOut"
                  },
                  "title": "Response List Characters Api Books  Book  Characters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Characters",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_character_api_books__book__characters_post",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Character",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/books/{book}/foreshadows": {
      "get": {
        "operationId": "list_foreshadows_api_books__book__foreshadows_get",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "open",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ForeshadowOut"
                  },
                  "title": "Response List Foreshadows Api Books  Book  Foreshadows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Foreshadows",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_foreshadow_api_books__book__foreshadows_post",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForeshadowInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForeshadowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Foreshadow",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/books/{book}/world-entries": {
      "get": {
        "operationId": "list_world_entries_api_books__book__world_entries_get",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "parent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorldEntryOut"
                  },
                  "title": "Response List World Entries Api Books  Book  World Entries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List World Entries",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_world_entry_api_books__book__world_entries_post",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorldEntryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldEntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create World Entry",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/characters/{char_id}": {
      "delete": {
        "operationId": "delete_character_api_characters__char_id__delete",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Char Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Character",
        "tags": [
          "settings"
        ]
      },
      "get": {
        "operationId": "get_character_api_characters__char_id__get",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Char Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "update_character_api_characters__char_id__patch",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Char Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Character",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/characters/{char_id}/affected-chapters": {
      "get": {
        "operationId": "affected_chapters_api_characters__char_id__affected_chapters_get",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "char_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Char Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AffectedChapter"
                  },
                  "title": "Response Affected Chapters Api Characters  Char Id  Affected Chapters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Affected Chapters",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/foreshadows/{fs_id}": {
      "patch": {
        "operationId": "update_foreshadow_api_foreshadows__fs_id__patch",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "fs_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Fs Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForeshadowUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForeshadowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Foreshadow",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/world-entries/{entry_id}": {
      "delete": {
        "operationId": "delete_world_entry_api_world_entries__entry_id__delete",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete World Entry",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "update_world_entry_api_world_entries__entry_id__patch",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorldEntryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldEntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update World Entry",
        "tags": [
          "settings"
        ]
      }
    }
  }
}